Use 64 bit integer (long or System.Int64) instead of 32 bit integer (int)

+++ NetStandardLibrary/Helper.cs
@@ -5,8 +5,8 @@ namespace NetStandardLibrary
     public class Helper
     {
         public static DateTime GetDate() => DateTime.Now;
-        private int? number;
+        private long? number;
 
-        public int? Number { get => number; set => number = value; }
+        public long? Number { get => number; set => number = value; }
     }
 }
